+++ packages/orq-rc/src/models/operations/createconversation.ts
@@ -270,7 +270,7 @@ export const CreateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfb68tc18dyv3hsncb31kyjn"),
+  _id: z.string().default("conv_01kfbbz05kaqhs7atv1wtqmnec"),
   kind: CreateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -302,7 +302,7 @@ export const CreateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfb68tc18dyv3hsncb31kyjn"),
+  id: z.string().default("conv_01kfbbz05kaqhs7atv1wtqmnec"),
   kind: CreateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),