+++ packages/orq-rc/src/models/operations/createdatasource.ts
@@ -576,7 +576,7 @@ export const CreateDatasourceResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("01KFB68V40T0RK9MFBQC6WVXK9"),
+  _id: z.string().default("01KFBBZ0NCS5FDQ024X0X9222X"),
   display_name: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$inboundSchema,
@@ -619,7 +619,7 @@ export const CreateDatasourceResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   CreateDatasourceResponseBody
 > = z.object({
-  id: z.string().default("01KFB68V40T0RK9MFBQC6WVXK9"),
+  id: z.string().default("01KFBBZ0NCS5FDQ024X0X9222X"),
   displayName: z.string(),
   description: z.string().optional(),
   status: CreateDatasourceStatus$outboundSchema,