+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -251,7 +251,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfb68tc3m6ev51g6xh572zgq"),
+  _id: z.string().default("conv_01kfbbz05qeaq2jt4aq9cwqs2n"),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -283,7 +283,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfb68tc3m6ev51g6xh572zgq"),
+  id: z.string().default("conv_01kfbbz05qeaq2jt4aq9cwqs2n"),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),