+++ packages/orq-rc/src/models/operations/updateconversation.ts
@@ -316,7 +316,7 @@ export const UpdateConversationResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kfb68tcbpv2tew7tspv2tave"),
+  _id: z.string().default("conv_01kfbbz05ztt4e5k6n2k26anmt"),
   kind: UpdateConversationKind$inboundSchema,
   displayName: z.string(),
   createdAt: z.number(),
@@ -348,7 +348,7 @@ export const UpdateConversationResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   UpdateConversationResponseBody
 > = z.object({
-  id: z.string().default("conv_01kfb68tcbpv2tew7tspv2tave"),
+  id: z.string().default("conv_01kfbbz05ztt4e5k6n2k26anmt"),
   kind: UpdateConversationKind$outboundSchema,
   displayName: z.string(),
   createdAt: z.number(),